✨ 선호 장소 뷰 구현
선호 장소 뷰 구현

+++ Catchy/Catchy/Model/Response/PlaceCourse/PlaceDetailResponse.swift
@@ -32,7 +32,7 @@ struct PlaceDetailResponse: Codable, Hashable {
     let activeTime: String
     
     /// 장소 평점
-    let rating: Float
+    let rating: Double
     
     /// 누군가 방문한 장소인지?
     let isVisited: Bool
